Move ScriptEngineLoader and getScriptEngine out of ScriptingMojo
Both of these are required by the JavadocMojo as well so extract these into utilities classes to avoid code duplication.
